refactor: Inline tests
